+++ eslint.config.js
@@ -14,6 +14,7 @@ export default antfu(
       '**/assets/**',
       '**/*.d.ts',
       '**/*.timestamp-*',
+      '**/test-reuslts',
       'test/unit/src/self',
       'test/unit/test/mocking/already-hoisted.test.ts',
       'test/cache/cache/.vitest-base/results.json',
